Make grid responsive
Why:

Currently, each cell has a fixed width and height. Instead, we would
like the grid to be response to the viewport, and for each cell to have
a width and height based on the width and height of the grid.

This commit:

Makes the grid responsive

Updates the design of the color picker

Makes some minor refactors